requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
boto3==1.28.79
botocore==1.31.83
celery==5.3.5
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.33.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.1
django-celery-results==2.5.0
django-cloudinary-storage==0.3.0
django-configurations==2.2
django-cors-headers==3.13.0
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-extra-fields==3.0.2
django-filter==23.2
django-hashids==0.7.0
django-ratelimit==4.1.0
django-storages==1.13.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.4
gunicorn==21.2.0
hashids==1.3.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.3.3
mailchimp-transactional==1.0.50
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
prompt-toolkit==3.0.40
psycopg2==2.9.6
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
s3transfer==0.7.0
sentry-sdk==1.31.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.1.0
wcwidth==0.2.9
whitenoise==6.5.0
